Move page layout component
refs #100
